+++ packages/react/src/components/Box.tsx
@@ -15,7 +15,7 @@ export function Box(props: BoxProps) {
     {/* className="flex flex-col gap-2" */}
       {props.as && (
         <props.as className={`${props.className} p-6 rounded-md bg-gray-800 border-solid border-[1px] border-gray-600`}>
-          {props.content && <Text content={props.content}></Text>}
+          {props.content && <Text>{props.content}</Text>}
           {Children.toArray(props.children).map((child, index) => (
             <div key={index}>{child}</div>
           ))}