+++ packages/react/src/components/Heading.tsx
@@ -2,8 +2,7 @@ import { ElementType } from "react";
 import "../styles/twbuild.css";
 import { clsx } from "clsx";
 
-export interface HeadingProps {
-  content: string;
+export interface HeadingProps extends React.HTMLAttributes<HTMLHeadingElement> {
   size?:
     | "md"
     | "lg"
@@ -16,6 +15,7 @@ export interface HeadingProps {
     | "8xl"
     | "9xl";
   as?: ElementType;
+  children?: React.ReactNode;
 }
 
 export function Heading(props: HeadingProps) {
@@ -35,15 +35,15 @@ export function Heading(props: HeadingProps) {
     <>
       {props.as ? (
         <props.as
-          className={`${classSize} font-default leading-shorter margin-0 text-gray-100`}
+          className={`${props.className} ${classSize} margin-0 leading-shorter text-gray-100`}
         >
-          {props.content}
+          {props.children}
         </props.as>
       ) : (
         <h2
-          className={`${classSize} font-default leading-shorter margin-0 text-gray-100`}
+          className={`${props.className} ${classSize} margin-0 leading-shorter text-gray-100`}
         >
-          {props.content}
+          {props.children}
         </h2>
       )}
     </>