+++ packages/react/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @isaias-ui/react
 
+## 1.3.0
+
+### Minor Changes
+
+- Box/Heading/Text: Use children instead of content prop
+  Updates other components to pass children
+  Updates button tertiary variant design
+
+## 1.2.3
+
+### Patch Changes
+
+- minor fixes for text-input
+
+## 1.2.2
+
+### Patch Changes
+
+- Heading component: extends html props and pass className down
+
+## 1.2.1
+
+### Patch Changes
+
+- Update the button component to receive child instead of a prop
+
 ## 1.2.0
 
 ### Minor Changes